Enable public API by default
Docker runs as a resident VM on PCs in most cases. Therefore
we should offer public API as open by default since the risk
of the machine getting attacked is minimal as the machine should
be firewalled and safe. This saves many users the hurdle of
changing the config themselves as well.